k8s-airflow/requirements.txt
pandas == 1
numpy == 1.21.6
apache-airflow-providers-amazon == 7.1.0
apache-airflow-providers-snowflake== 4.0.2